requirements.txt
pandas
kaggle
python-dotenv
snowflake-connector-python
dbt-core
dbt-snowflake
